# expression: implement vectorized evaluation for 'builtinGTIntSig'