… expression/builtin_compare_vec_test.go