<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for builtinGTIntSig, for #12103

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-VecBuiltinFunc-8         	  349441	      3238 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-NonVecBuiltinFunc-8      	   27792	     42825 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-VecBuiltinFunc#01-8      	  381259	      3128 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-NonVecBuiltinFunc#01-8   	   29038	     41925 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-VecBuiltinFunc#02-8      	  321118	      3711 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-NonVecBuiltinFunc#02-8   	   27728	     41709 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-VecBuiltinFunc#03-8      	  349153	      3407 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinCompareFunc/builtinGTIntSig-NonVecBuiltinFunc#03-8   	   28828	     42035 ns/op	       0 B/op	       0 allocs/op
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test